% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_operations.R
\name{add_parent_to_loop}
\alias{add_parent_to_loop}
\title{Add rows of the parent dataframe}
\usage{
add_parent_to_loop(loop, parent, variables.to.keep = NULL,
  uuid.name.loop = NULL, uuid.name.parent = NULL)
}
\arguments{
\item{loop}{a dataframe containing the loops}

\item{parent}{a dataframe containing the parent informations}

\item{variables.to.keep}{optional: a string or a vector of string containing the parent columns names that need to be copy in the loop dataset}

\item{uuid.name.loop}{optional: Specify the loop column containing the uuids. If not specify, searches for column containing uuid string}

\item{uuid.name.parent}{optional: Specify the parent column containing the uuids. If not specify, searches for column containing uuid string}
}
\value{
loop dataframe with the chosen parent variables in addition
}
\description{
Add rows of the parent dataframe
}
\details{
Searches for parent row for each row in loop, select the rows of the parent dataframe and add them to the loop dataframe
}
\examples{
parent <- data.frame(uuid=1:10, age=sample(10,30,60),gender=sample(c("F","M"),10,replace = T) )
child <- data.frame (parent_uuid=sample(1:10,20,replace = T), age=sample(20,1,18)  gender=sample(c("F","M"),20,replace = T))
family<- add_parent_to_loop(child,parent,c("age"))
equivalente to
family<- add_parent_to_loop(child,parent,"parent_uuid","uuid",c("age"))
}
